style: 앱설정 / 시간 설정 레이아웃
